classifier/requirements.txt
transformers
pillow
tqdm
chromadb
aio-pika
pydantic
pydantic-settings
psycopg2-binary
sqlalchemy
pycocotools
torchmetrics
chromadb
supervision
boto3